combine the two test files
